Created Info Window component
