+++ application/src/app/components/MapMarker.tsx
@@ -6,6 +6,7 @@ import X from './icons/X';
 import Rating from './Rating';
 import { formatCategory } from '../utils/utils';
 import { LocationType } from '@prisma/client';
+import MapInfoWindow from './MapInfoWindow';
 
 type MapMarkerStyle = "default" | "bold" | "filled";
 
@@ -23,11 +24,6 @@ const mapPinBoldEdited =
   'M 5.0263818e-7,-184.5 A 44,44 0 1 0 44,-140.5 44.05,44.05 0 0 0 5.0263818e-7,-184.5 Z m 0,64 A 20,20 0 1 1 20,-140.5 20,20 0 0 1 5.0263818e-7,-120.5 Z m 0,-112 A 92.1,92.1 0 0 0 -92,-140.5 c 0,77.360001 81.640001,135.4000009 85.1200005,137.8300009 a 12,12 0 0 0 13.76,0 A 259,259 0 0 0 49.06,-41.669999 C 77.15,-73.929999 92,-108.13 92,-140.5 A 92.1,92.1 0 0 0 5.0263818e-7,-232.5 Z M 31.3,-57.789999 A 249.35,249.35 0 0 1 5.0263818e-7,-27.609999 249.35,249.35 0 0 1 -31.3,-57.789999 C -48,-77.129999 -68,-107.19 -68,-140.5 a 68,68 0 0 1 136,0 c 0,33.31 -20,63.370001 -36.7,82.710001 z';
 const mapPinFilledEdited =
   'm -0.00506667,-224.2533 a 88.1,88.1 0 0 0 -88.00000033,88 c 0,75.3 80.0000003,132.1700001 83.4100003,134.5500001 a 8,8 0 0 0 9.18,0 c 3.41,-2.38 83.4099997,-59.2500001 83.4099997,-134.5500001 a 88.1,88.1 0 0 0 -87.99999967,-88 z m 0,56 a 32,32 0 1 1 -32.00000033,32 32,32 0 0 1 32.00000033,-32 z';
-
-const infoWindowOptions: google.maps.InfoWindowOptions = {
-  maxWidth: 300,
-  headerDisabled: true,
-};
 
 const getSvgPath = (style: MapMarkerStyle): string => {
   if (style == "default") {
@@ -97,7 +93,8 @@ const MapMarker = ({ location, selectedMarkerId, setSelectedMarkerId, style }: M
     >
       {/* Style 2: Imgae on the side */}
       {/* Style 1: Image on the top */}
-      {selectedMarkerId === location.id && <InfoWindow options={infoWindowOptions}>
+      {selectedMarkerId === location.id && <MapInfoWindow location={location} />}
+      {/* {selectedMarkerId === location.id && <InfoWindow options={infoWindowOptions}>
         
         <div
           className="flex flex-col items-center w-[300px] h-full p-3 relative rounded-lg border-4 border-[#0000FF]"
@@ -133,63 +130,63 @@ const MapMarker = ({ location, selectedMarkerId, setSelectedMarkerId, style }: M
             />
           </div>
         </div>
-      </InfoWindow>}
+      </InfoWindow>} */}
     </Marker>
   );
 
   {
     /* V2: InfoWindow closes and opens based on hover actions*/
   }
-  return (
-    <Marker
-      key={location.id}
-      position={{ lat: location.latitude, lng: location.longitude }}
-      icon={{
-        path: mapPinFilledEdited,
-        fillColor: '#d1daaf',
-        fillOpacity: 1.0,
-        scale: 0.15,
-        strokeColor: '#869747',
-        strokeWeight: 4,
-        anchor: new google.maps.Point(0, 0),
-      }}
-      onMouseOver={() => setShowInfo(true)}
-      onMouseOut={() =>
-        setTimeout(() => {
-          if (!hoveringInfoWindow) {
-            setShowInfo(false);
-          }
-        }, 300)
-      }
-    >
-      {showInfo && (
-        <InfoWindow options={infoWindowOptions}>
-          <div
-            className="flex flex-col items-center w-full h-full"
-            onMouseEnter={() => {
-              setHoveringInfoWindow(true);
-            }}
-            onMouseLeave={() => {
-              // setShowInfo(false);
-              setHoveringInfoWindow(false);
-            }}
-          >
-            <div className="relative w-full h-[150px]">
-              <Image
-                className="object-cover rounded-lg bg-slate"
-                alt={location.name}
-                fill={true}
-                src={location.imageWebLink}
-              />
-            </div>
-            <div className="font-shantell text-lg mb-2 text-center mt-2">
-              {location.name}
-            </div>
-          </div>
-        </InfoWindow>
-      )}
-    </Marker>
-  );
+  // return (
+  //   <Marker
+  //     key={location.id}
+  //     position={{ lat: location.latitude, lng: location.longitude }}
+  //     icon={{
+  //       path: mapPinFilledEdited,
+  //       fillColor: '#d1daaf',
+  //       fillOpacity: 1.0,
+  //       scale: 0.15,
+  //       strokeColor: '#869747',
+  //       strokeWeight: 4,
+  //       anchor: new google.maps.Point(0, 0),
+  //     }}
+  //     onMouseOver={() => setShowInfo(true)}
+  //     onMouseOut={() =>
+  //       setTimeout(() => {
+  //         if (!hoveringInfoWindow) {
+  //           setShowInfo(false);
+  //         }
+  //       }, 300)
+  //     }
+  //   >
+  //     {showInfo && (
+  //       <InfoWindow options={infoWindowOptions}>
+  //         <div
+  //           className="flex flex-col items-center w-full h-full"
+  //           onMouseEnter={() => {
+  //             setHoveringInfoWindow(true);
+  //           }}
+  //           onMouseLeave={() => {
+  //             // setShowInfo(false);
+  //             setHoveringInfoWindow(false);
+  //           }}
+  //         >
+  //           <div className="relative w-full h-[150px]">
+  //             <Image
+  //               className="object-cover rounded-lg bg-slate"
+  //               alt={location.name}
+  //               fill={true}
+  //               src={location.imageWebLink}
+  //             />
+  //           </div>
+  //           <div className="font-shantell text-lg mb-2 text-center mt-2">
+  //             {location.name}
+  //           </div>
+  //         </div>
+  //       </InfoWindow>
+  //     )}
+  //   </Marker>
+  // );
 };
 
 export default MapMarker;
